Fix PS1 to be more usable
I don't care who I'm logged in as (99.99% of the time it will be my own account)
so bets remove it and see more of the path I am operating on.

This also tells vim to automatically reload should files change outside of the
current buffer so I'm always on the latest.